gather SchemaTable sub-components, styles per component
Like TableLayout, the table cells are very specific to their usage in a
SchemaTable, so let's put them in the same directory.

Furthermore, let's attach styles to the component where they are used.
