+++ src/components/SchemaTable/LeftCell.jsx
@@ -19,23 +19,73 @@ import {
   TOOLTIP_DESCRIPTIONS,
 } from '../../utils/constants';
 
-/**
- * Dynamically generate styles for indentations to be used for
- * displaying the data structure of the schemas.
- */
 const useStyles = makeStyles(theme => ({
+  /**
+   * Dynamically generate styles for indentations to be used for
+   * displaying the data structure of the schemas.
+   */
   indentation: {
     marginLeft: indent => theme.spacing(indent * 2),
   },
+
+  /**
+   * Typography within the cell
+   */
+  typography: {
+    color: theme.palette.text.primary,
+    display: 'flex',
+    alignItems: 'center',
+    height: theme.spacing(4.5),
+  },
+
+  /**
+   * Name text displayed within a LeftCell.
+   */
+  name: {
+    marginRight: theme.spacing(0.5),
+  },
+  /**
+   * Highlight the type for the schema or sub-schema displayed
+   * within a LeftCell.
+   */
+  code: {
+    backgroundColor: theme.palette.text.primary,
+    color: theme.palette.getContrastText(theme.palette.text.primary),
+    padding: `0 ${theme.spacing(0.5)}px`,
+    fontSize: theme.typography.subtitle2.fontSize,
+    fontWeight: theme.typography.subtitle2.fontWeight,
+    fontFamily: theme.typography.subtitle2.fontFamily,
+  },
+  /**
+   * Warning icon to inform missing type in LeftCell.
+   */
+  missingType: {
+    display: 'flex',
+    alignItems: 'center',
+  },
+  /**
+   * Comments used for combination types in LeftCell.
+   */
+  comment: {
+    color: theme.palette.text.hint,
+  },
+  /**
+   * Prefixes used to notate special properties of data types in
+   * lines of LeftCell. (ex. 'required', 'contains' keywords)
+   */
+  prefix: {
+    color: theme.palette.error.main,
+    padding: `0 ${theme.spacing(0.5)}px`,
+  },
+  /**
+   * Button used to expand or shrink a $ref.
+   */
+  refButton: {
+    marginLeft: theme.spacing(1),
+  },
 }));
 
-function NormalLeftRow({
-  classes,
-  treeNode,
-  refType,
-  setSchemaTree,
-  references,
-}) {
+function LeftCell({ treeNode, refType, setSchemaTree, references }) {
   const { schema, path } = treeNode;
   const schemaType = schema._type;
   const name = schema._name;
@@ -44,7 +94,7 @@ function NormalLeftRow({
    * (length of path = depth of the current treeNode)
    */
   const indentSize = path.length;
-  const styles = useStyles(indentSize);
+  const classes = useStyles(indentSize);
   /**
    * Create a text for the name of the schema.
    */
@@ -192,7 +242,7 @@ function NormalLeftRow({
       <Typography
         component="div"
         variant="subtitle2"
-        className={classNames(classes.line, styles.indentation)}>
+        className={classNames(classes.typography, classes.indentation)}>
         {name && nameText}
         {typeSymbol}
         {requiredMark}
@@ -202,7 +252,7 @@ function NormalLeftRow({
   );
 }
 
-NormalLeftRow.propTypes = {
+LeftCell.propTypes = {
   /**
    * Style for rows and lines for left rows of the schema table.
    * (necessary to maintain consistency with right panel's rows and lines)
@@ -244,9 +294,9 @@ NormalLeftRow.propTypes = {
   references: object,
 };
 
-NormalLeftRow.defaultProps = {
+LeftCell.defaultProps = {
   setSchemaTree: NOOP,
   references: {},
 };
 
-export default React.memo(NormalLeftRow);
+export default React.memo(LeftCell);
